Реализация механики добавления в друзья в UserService Создание NotFoundException и его обработчика

+++ src/main/java/ru/yandex/practicum/filmorate/controller/UserController.java
@@ -40,7 +40,7 @@ public class UserController {
     }
 
     @DeleteMapping("/{userId}")
-    public void deleteUser(@PathVariable int userId) {
+    public void deleteUser(@PathVariable Long userId) {
         log.info("Запрос на удаление пользователя: {}", userId);
         inMemoryUserStorage.deleteUser(userId);
     }
